Respond to PR feedback
I've made it more explicit that the minikube Docker driver is required
for the demo. I have also added a row to the required software table
that includes Docker Desktop/Docker Engine. I've tested this on my
Linux desktop, and modified commands to support Docker Engine as well
as Docker Desktop (i.e., "minikube tunnel" exposes a private IP address
beside 127.0.0.1 for the load balancer).

Also made a couple of minor tweaks, and removed the mention of
localhost in relation to the Web UI.